AsyncGRPO LoRA: move the adapter helpers out of utils; test the save on the runner's device
Two review follow-ups on #7017.

`round_lora_rank`, `validate_lora_for_vllm_sync` and `save_lora_adapter` move
from `trl/trainer/utils.py` into `async_grpo_trainer.py`. Their only caller is
experimental, and main code should not exist to serve experimental code; they
get promoted back to utils when the GRPO follow-up gives them a second caller.
`shutil` and `Accelerator` were imported into utils for these three alone and
leave with them. The DTensor import stays: `patch_chunked_lm_head` uses it.

`TestSaveLoraAdapter`'s stub accelerator now reports `torch_device` instead of
a hardcoded `cpu`. The model stays on CPU, which is the `fsdp_offload_params`
shape: `save_lora_adapter` moves each shard to `accelerator.device` before the
gather and brings the result back to CPU before writing. With `cpu` hardcoded
both moves were no-ops, so a tensor left on the wrong device could never have
failed the test on a GPU runner. Now it is a real round trip there.
